Add MVVM model for Deck Cell
